+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/MainCarContext.kt
@@ -12,7 +12,7 @@ import com.mapbox.navigation.core.MapboxNavigationProvider
 import com.mapbox.navigation.core.formatter.MapboxDistanceFormatter
 import com.mapbox.navigation.ui.maneuver.api.MapboxManeuverApi
 import com.mapbox.navigation.utils.internal.JobControl
-import com.mapbox.search.SearchEngine
+import com.mapbox.search.SearchEngineSettings
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
@@ -21,9 +21,9 @@ import kotlinx.coroutines.SupervisorJob
 class MainCarContext(
     val carContext: CarContext,
     val mapboxCarMap: MapboxCarMap,
-    val searchEngine: SearchEngine,
     val feedbackPollProvider: CarFeedbackPollProvider = CarFeedbackPollProvider(),
     val routeOptionsInterceptor: CarRouteOptionsInterceptor = CarRouteOptionsInterceptor { it },
+    val searchEngineSettings: SearchEngineSettings? = null
 ) {
     val carSettingsStorage = CarSettingsStorage(carContext)