+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/car/MainCarSession.kt
@@ -29,10 +29,7 @@ import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.base.options.NavigationOptions
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
 import com.mapbox.navigation.core.trip.session.TripSessionState
-import com.mapbox.navigation.qa_test_app.car.search.CarSearchLocationProvider
 import com.mapbox.navigation.qa_test_app.utils.Utils
-import com.mapbox.search.MapboxSearchSdk
-import com.mapbox.search.SearchEngineSettings
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
@@ -52,6 +49,7 @@ class MainCarSession : Session() {
         logAndroidAuto("MainCarSession constructor")
         val logoSurfaceRenderer = CarLogoSurfaceRenderer()
         val compassSurfaceRenderer = CarCompassSurfaceRenderer()
+        MapboxNavigationApp.attach(lifecycleOwner = this)
         lifecycle.addObserver(object : DefaultLifecycleObserver {
 
             override fun onCreate(owner: LifecycleOwner) {
@@ -68,13 +66,7 @@ class MainCarSession : Session() {
 
                 mapboxCarMap = MapboxCarMap(MapInitOptions(context = carContext))
                 mapboxCarMap.registerObserver(carMapStyleLoader)
-                val searchEngine = MapboxSearchSdk.createSearchEngineWithBuiltInDataProviders(
-                    SearchEngineSettings(
-                        Utils.getMapboxAccessToken(carContext.applicationContext),
-                        MapboxNavigationApp.getObserver(CarSearchLocationProvider::class),
-                    ),
-                )
-                val mainCarContext = MainCarContext(carContext, mapboxCarMap, searchEngine)
+                val mainCarContext = MainCarContext(carContext, mapboxCarMap)
                     .also { mainCarContext = it }
                 val mainScreenManager = MainScreenManager(mainCarContext)
                     .also { mainScreenManager = it }
@@ -111,8 +103,6 @@ class MainCarSession : Session() {
                 MapboxNavigationApp.unregisterObserver(notificationInterceptor)
             }
         })
-
-        MapboxNavigationApp.attach(lifecycleOwner = this)
     }
 
     override fun onCreateScreen(intent: Intent): Screen {